Move components into seperate file
