+++ src/video/data/plot-data.ts
@@ -1,0 +1,14 @@
+export const dataset1 = [
+  [10, 12],
+  [17, 13],
+  [20, 10],
+  [20, 18],
+  [29, 28],
+  [32, 21],
+  [33, 26],
+  [38, 80],
+  [45, 26],
+  [47, 30],
+  [70, 70],
+  [95, 70],
+] as [number, number][]
